requirements.txt
asgiref==3.2.10
Django==3.1.1
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
importlib-metadata==2.0.0
Markdown==3.2.2
Pillow==7.2.0
psycopg2==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
style==1.1.0
update==0.0.1
zipp==3.3.0
